IDE compatibility and removed versioning inside smartapp, will use github tags instead
