+++ smartapps/opendash/open-dash.src/open-dash.groovy
@@ -15,14 +15,14 @@
 *  Author: Open-Dash
 *  based on https://github.com/jodyalbritton/apismartapp/blob/master/endpoint.groovy
 *  weather code from https://github.com/Dianoga/my-smartthings/blob/master/devicetypes/dianoga/weather-station.src/weather-station.groovy
-*
+*  
 *  To Donate to this project please visit https://open-dash.com/donate/
 */
 
 import groovy.json.JsonBuilder
 
 definition(
-    name: "Open-Dash ${appVersion()}",
+    name: "Open-Dash",
     namespace: "opendash",
     author: "Open-Dash",
     description: "Open-Dash",
@@ -30,24 +30,23 @@ definition(
     iconUrl: "https://s3.amazonaws.com/smartapp-icons/Convenience/Cat-Convenience.png",
     iconX2Url: "https://s3.amazonaws.com/smartapp-icons/Convenience/Cat-Convenience@2x.png",
 )
-def appVersion() {"0.0.6"}
 
 mappings {
     // location
     path("/locations") 							{	action: [	GET: "listLocation"        														]}
-    // modes
+    // modes    
     path("/modes") 								{   action: [   GET: "listModes"        														]}
     path("/modes/:id") 							{	action: [   GET: "switchMode"        														]}
     // hub
     path("/hubs") 								{   action: [   GET: "listHubs"		       														]}
     path("/hubs/:id") 							{   action: [   GET: "getHubDetail"        														]}
-    // devices
+    // devices  
     path("/devices") 							{   action: [   GET: "listDevices"        														]}
     path("/devices/:id") 						{  	action: [   GET: "listDevices"        														]}
     path("/devices/:id/events") 				{   action: [   GET: "listDeviceEvents"        													]}
-    path("/devices/:id/commands") 				{	action: [	GET: "listDeviceCommands"        												]}
-    path("/devices/:id/:command")				{   action: [	GET: "sendDeviceCommand"          												]}
-    path("/devices/:id/:command/:secondary")	{   action: [   GET: "sendDeviceCommandSecondary"           									]}
+    path("/devices/:id/commands") 				{	action: [	GET: "listDeviceCommands"        												]}    
+    path("/devices/:id/:command")				{   action: [	GET: "sendDeviceCommand"          												]}    
+    path("/devices/:id/:command/:secondary")	{   action: [   GET: "sendDeviceCommandSecondary"           									]}   
 
     // Routines
     path("/routines") 							{   action: [   GET: "listRoutines"        														]}
@@ -55,12 +54,11 @@ mappings {
     path("/updates") 							{   action: [   GET: "updates"        															]}
     path("/allDevices") 						{   action: [   GET: "allDevices"        														]}
     path("/devicetypes")						{	action: [ 	GET: "listDeviceTypes" 															]}
-    path("/version")							{	action: [ 	GET: "getVersion" 																]}
     path("/weather")							{	action: [ 	GET: "getWeather" 																]}
 }
 
 private def getCapabilities() {
-    [   //Capability Prefrence Reference			Display Name					Subscribed Name						Subscribe Attribute
+    [   //Capability Prefrence Reference			Display Name					Subscribed Name						Subscribe Attribute        
         ["capability.accelerationSensor",			"Accelaration Sensor",			"accelerations",					"acceleration"				],
         ["capability.actuator",						"Actuator",						"actuators",						""							],
         ["capability.alarm",						"Alarm",						"alarms",							"alarm"						],
@@ -122,7 +120,7 @@ private def getCapabilities() {
         ["capability.voltageMeasurement",			"Voltage Measurement",			"voltageMeasurements",				"voltage"					],
         ["capability.waterSensor",					"Water Sensor",					"waterSensors",						"water"						],
         ["capability.windowShade",					"Window Shade",					"windowShades",						"windowShade"				],
-    ]
+    ]  
 }
 
 preferences {
@@ -149,6 +147,19 @@ def initialize() {
             subscribe(settings[cap[2]], cap[3], handleEvent)
         }
     }
+    subscribe(location, "alarmSystemStatus", alarmHandler)  
+}
+
+/****************************
+* Alarm Methods
+****************************/
+
+def alarmHandler(evt) {
+	if (!state.updates) state.updates = []
+    //evt.value = ["stay","away","off"]
+    def shm = eventJson(evt)
+    shm.id = "shm"
+    state.updates << shm
 }
 
 /****************************
@@ -429,9 +440,10 @@ def allDevices() {
             deviceData << [ "attributes" : attributes ]
             def cmds = []
             i.supportedCommands?.each {
-            		cmds << it.name
+            		cmds << ["command" : it.name ]
         	}
-            deviceData << [ "commands" : cmds ]
+            //log.debug cmds
+            deviceData << [ "commands" : cmds ] //i.supportedCommands.toString() ]  //TODO fix this to parse to an object
             allAttributes << deviceData
         }
     }
@@ -444,9 +456,9 @@ def listDeviceTypes() {
     log.debug "${allSubscribed.size()} Unique Devices" // is $uniqueDevices"
 
     allSubscribed.each {
-        it.collect{ i ->
+        it.collect{ i ->    
             if (!deviceData.contains(i?.typeName)) {
-                deviceData << i?.typeName
+                deviceData << i?.typeName  
             }
         } //.flatten().unique { it }
     }
@@ -454,9 +466,7 @@ def listDeviceTypes() {
 }
 
 /* General API Functions */
-def getVersion() {
-    render contentType: "text/json", data: appVersion();
-}
+
 
 /* WebHook API Call on Subscribed Change */
 private logField(evt, Closure c) {
@@ -537,7 +547,7 @@ def getWeather() {
     log.debug obs
     log.debug obs.alertString
     if (obs) {
-       return obs
+       return obs  
     }
 }
 
@@ -567,7 +577,7 @@ log.debug evt.user // null?
 
     if(x) {
         for(i in x) {
-            state.updates.remove(i)
+            state.updates.remove(i) 
         }
     }
     state.updates << js
@@ -575,17 +585,17 @@ log.debug evt.user // null?
 
 private getAllSubscribed() {
     def dev_list = []
-    capabilities.each {
-        dev_list << settings[it[2]]
+    capabilities.each { 
+        dev_list << settings[it[2]] 
     }
     return dev_list?.findAll()?.flatten().unique { it.id }
 }
 
 private item(device, s) {
-    device && s ? [device_id: device.id,
-                   label: device.displayName,
-                   name: s.name, value: s.value,
-                   date: s.date, stateChange: s.stateChange,
+    device && s ? [device_id: device.id, 
+                   label: device.displayName, 
+                   name: s.name, value: s.value, 
+                   date: s.date, stateChange: s.stateChange, 
                    eventSource: s.eventSource] : null
 }
 
@@ -682,12 +692,9 @@ private estimateLux(sunriseDate, sunsetDate, weatherIcon) {
 }
 
 
-//TODO  Add location subscription subscribe(location, "alarmSystemStatus", alarmStatus) for alarmStatus function to store alarm state
 //TODO get SHM status String alarmSystemStatus = "${location?.currentState("alarmSystemStatus").stringValue}"
 //TODO create alarmStatus function for processing changes to alarm state aka SHM
 //TODO update alarm state with this sendLocationEvent(name: "alarmSystemStatus", value: status)  values = off,away,stay
 //TODO add commands and attributes defn, maybe liberate from https://github.com/ady624/webCoRE/blob/5f41cdcaf08616fb021021f7a4a4b3ccb1b7e239/smartapps/ady624/webcore.src/webcore.groovy
 //TODO add function and endpoint for sending notifications
 //TODO add endpoint for controlling a group of deviceids in JSON (for group off commands)
-//TODO add device health endpoint and nightly scheduled function to check device history for last activity in 24 hours, create offline list, and endpoint that checks those devices for activity since last run
-//TODO add UI info for inside ST for donations and other support related info
